feat: implement GET API call for fetching ChannelData

+++ app/src/main/java/com/sandor/kis/tvguide/data/model/Tstv.kt
@@ -1,8 +1,8 @@
 package com.sandor.kis.tvguide.data.model
 
 data class Tstv(
-    val availabilityEnd: Int,
-    val availabilityStart: Int,
+    val availabilityEnd: Long,
+    val availabilityStart: Long,
     val mediaAvailable: Boolean,
     val mediaLocation: String,
     val sid: Int
